pyproject.toml
[tool.poetry]
name = "githubctl"
version = "0.1.0"
description = ""
authors = ["hasosyn <hasan.r.karasu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
typer = {extras = ["all"], version = "^0.12.3"}
urllib3 = "1.26.6"
requests = "^2.31.0"
python-dotenv = "^1.0.1"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
